+++ src/lib.rs
@@ -7,14 +7,15 @@ pub mod presentation;
 pub mod services;
 use actor_registry::ActorRegistry;
 use anyhow::Result;
+use domains::IoError;
 use domains::append_only_files::interfaces::TAof;
 use domains::caches::cache_manager::CacheManager;
+use domains::cluster_actors::ClusterActor;
 use domains::cluster_actors::commands::ClusterCommand;
-use domains::cluster_actors::replication::IS_LEADER_MODE;
+use domains::cluster_actors::replication::{IS_LEADER_MODE, ReplicationInfo};
 use domains::config_actors::config_manager::ConfigManager;
 use domains::saves::snapshot::snapshot_applier::SnapshotApplier;
 use domains::saves::snapshot::snapshot_loader::SnapshotLoader;
-use domains::IoError;
 pub use init::Environment;
 use presentation::clients::ClientController;
 use presentation::cluster_in::communication_manager::ClusterCommunicationManager;
@@ -42,15 +43,13 @@ impl StartUpFacade {
         let snapshot_applier =
             SnapshotApplier::new(cache_manager.clone(), config_manager.startup_time);
 
-        let cluster_actor_handler = ClusterCommunicationManager::run(
-            notifier,
+        let cluster_actor_handler = ClusterActor::run(
             env.ttl_mills,
             env.hf_mills,
-            env.replicaof,
-            env.host.clone(),
-            env.port,
-            aof,
+            ReplicationInfo::new(env.replicaof, &env.host, env.port),
             cache_manager.clone(),
+            notifier,
+            aof,
         );
 
         let registry = ActorRegistry {
