+++ src/presentation/cluster_in/communication_manager.rs
@@ -1,14 +1,12 @@
 use crate::{
     domains::{
-        append_only_files::interfaces::TAof,
-        caches::cache_manager::CacheManager,
-        cluster_actors::{ClusterActor, commands::ClusterCommand, replication::ReplicationInfo},
+        cluster_actors::{commands::ClusterCommand, replication::ReplicationInfo},
         peers::identifier::PeerIdentifier,
     },
     make_smart_pointer,
 };
-use std::time::Duration;
-use tokio::{sync::mpsc::Sender, time::interval};
+
+use tokio::sync::mpsc::Sender;
 
 #[derive(Clone)]
 pub struct ClusterCommunicationManager(pub(crate) Sender<ClusterCommand>);
@@ -16,43 +14,6 @@ pub struct ClusterCommunicationManager(pub(crate) Sender<ClusterCommand>);
 make_smart_pointer!(ClusterCommunicationManager, Sender<ClusterCommand>);
 
 impl ClusterCommunicationManager {
-    pub fn run(
-        notifier: tokio::sync::watch::Sender<bool>,
-        node_timeout: u128,
-        heartbeat_fq_mills: u64,
-        replicaof: Option<(String, String)>,
-        host: String,
-        port: u16,
-        aof: impl TAof,
-        cache_manager: CacheManager,
-    ) -> Sender<ClusterCommand> {
-        let (actor_handler, cluster_message_listener) = tokio::sync::mpsc::channel(100);
-
-        tokio::spawn(
-            ClusterActor::new(
-                node_timeout,
-                ReplicationInfo::new(replicaof, &host, port),
-                cache_manager,
-                notifier,
-            )
-            .handle(cluster_message_listener, aof),
-        );
-
-        tokio::spawn({
-            let heartbeat_sender = actor_handler.clone();
-            let mut heartbeat_interval = interval(Duration::from_millis(heartbeat_fq_mills));
-            async move {
-                loop {
-                    heartbeat_interval.tick().await;
-                    let _ = heartbeat_sender.send(ClusterCommand::SendHeartBeat).await;
-                }
-            }
-        });
-
-        // TODO peer state may need to be picked up from a persistent storage on restart case
-        actor_handler
-    }
-
     pub(crate) async fn get_peers(&self) -> anyhow::Result<Vec<PeerIdentifier>> {
         let (tx, rx) = tokio::sync::oneshot::channel();
         self.send(ClusterCommand::GetPeers(tx)).await?;
